requirements.txt
numpy==1.19.1
opencv-python==4.4.0.46
Pillow==8.1.0

matplotlib~=3.3.3